[ssk122]: Initial restructuring of front-end files
