+++ react-front-end/src/index.js
@@ -3,6 +3,9 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import App from './App';
 import './index.css';
+// avoid using index.css split it into per page css
+// e.g. landing.css, finances.css, etc
+
 
 ReactDOM.render(
   <React.StrictMode>
